Split system purpose tests to a separate file
The main test file is getting a bit too big, let's split
system purpose helper module tests to a separate file.

Related: rhbz#1951709
